Add files via upload
This Project is based on Linear Regression , with the help of linear regression favorable outcomes of prediction is done significantly.
There various functions was used like Sklearn , Matplotlib , Seaborn , SVM and etc.
Here i had done data analysis of Gujarat Titans Team by taking data of 10 years.
OBJECTIVE OF THE PROJECT :-
This project is very helpful for the Recruiters to analyse the Team Performance and as well as Players Performance. So, that they can have strong and clear identifications of all all players. It is helpful for the Audience as well to rectify Players Score Performance.